requirements.txt

selenium
webdriver-manager
python-dotenv
twilio
schedule
